+++ src/fast_model/mod.rs
@@ -84,7 +84,7 @@ pub use resolve::*;
 pub use mesh_generate::{
     booleans_meshes_in_db, gen_inst_meshes, gen_meshes_in_db, process_meshes_update_db,
     process_meshes_update_db_deep, process_meshes_update_db_deep_default,
-    update_inst_relate_aabbs_by_refnos,
+    process_meshes_bran, update_inst_relate_aabbs_by_refnos,
 };
 
 pub const SEND_INST_SIZE: usize = 500;